Refactor redux store folder
