+++ packages/bierzo-wallet/src/store/index.ts
@@ -1,5 +1,17 @@
-import { applyMiddleware, compose, createStore, Middleware, Store } from 'redux';
-import reducer, { RootReducer } from './reducers';
+import { applyMiddleware, combineReducers, compose, createStore, Middleware, Reducer, Store } from 'redux';
+import { StateType } from 'typesafe-actions';
+import { extensionReducer, ExtensionState } from './extension';
+
+export interface RootReducer {
+  extension: ExtensionState;
+}
+
+const createRootReducer = (): Reducer<RootReducer> =>
+  combineReducers({
+    extension: extensionReducer,
+  });
+
+export type RootState = StateType<ReturnType<typeof createRootReducer>>;
 
 const composeEnhancers =
   (typeof window === 'object' && (window as any).__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) || // eslint-disable-line
@@ -8,7 +20,7 @@ const composeEnhancers =
 const middlewares: ReadonlyArray<Middleware> = [];
 
 export const configureStore = (): Store<RootReducer> => {
-  const store = createStore(reducer, composeEnhancers(applyMiddleware(...middlewares)));
+  const store = createStore(createRootReducer(), composeEnhancers(applyMiddleware(...middlewares)));
 
   if (process.env.NODE_ENV !== 'production' && module.hot) {
     module.hot.accept(
