--- packages/bierzo-wallet/src/store/reducers/index.ts
@@ -1,15 +1,0 @@
-import { combineReducers, Reducer } from 'redux';
-import { StateType } from 'typesafe-actions';
-import { extensionReducer, ExtensionState } from './extension';
-
-export interface RootReducer {
-  extension: ExtensionState;
-}
-
-const createRootReducer = (): Reducer<RootReducer> =>
-  combineReducers({
-    extension: extensionReducer,
-  });
-
-export type RootState = StateType<ReturnType<typeof createRootReducer>>;
-export default createRootReducer();
